feat(travel_explorer_training): add travel_rating package
This commit introduces a new package, `travel_rating`, to the project.  The package contains widgets and utilities for displaying and interacting with travel ratings.  This improves the user experience by providing more sophisticated rating components.  The `travel_rating` package is added as a dependency to the `travel_explorer_app`.  This allows the app to utilize the new rating components.